Save ProfileCard to separate file, tidy up
