+++ client/package.json
@@ -20,6 +20,8 @@
     "date-fns": "^2.22.1",
     "eslint": "^7.26.0",
     "formik": "^2.2.6",
+    "moment": "^2.29.1",
+    "moment-range": "^4.0.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
